requirements.txt
python-telegram-bot==20.7
Pillow==9.5.0
pytesseract==0.3.10
requests==2.31.0
